# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoltIQ",
    "description": "VoltIQ is an enterprise-grade AI-powered fleet electrification and asset intelligence platform.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/battery/predict": {
      "post": {
        "tags": [
          "Battery APM"
        ],
        "summary": "Predict Battery Health",
        "description": "Predict State of Health (SOH) and Remaining Useful Life (RUL) of an EV battery\nbased on incoming cell telemetry metrics (Voltage, Cycle Number, Temp, Capacity).",
        "operationId": "predict_battery_health_api_battery_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatteryTelemetryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatteryPredictionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/battery/rul": {
      "post": {
        "tags": [
          "Battery APM"
        ],
        "summary": "Predict Battery Rul",
        "description": "Predict Remaining Useful Life (RUL) only, using State of Health as input feature.",
        "operationId": "predict_battery_rul_api_battery_rul_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatteryRULInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatteryRULOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/carbon/analysis": {
      "post": {
        "tags": [
          "Carbon Intelligence Tracker"
        ],
        "summary": "Analyze Vehicle Carbon",
        "description": "Perform a comparative carbon emissions analysis for a specific vehicle (baseline vs EV scenario)\nover a custom annual distance and fuel type.",
        "operationId": "analyze_vehicle_carbon_api_carbon_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarbonAnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarbonAnalysisOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/query": {
      "post": {
        "tags": [
          "AI Fleet Advisor"
        ],
        "summary": "Query Fleet Advisor",
        "description": "Query the AI Fleet Advisor to get insights about battery telemetry, \nreadiness schedules, or carbon calculations.",
        "operationId": "query_fleet_advisor_api_chat_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatQueryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/battery/predict": {
      "post": {
        "tags": [
          "Battery APM"
        ],
        "summary": "Predict Battery Health",
        "description": "Predict State of Health (SOH) and Remaining Useful Life (RUL) of an EV battery\nbased on incoming cell telemetry metrics (Voltage, Cycle Number, Temp, Capacity).",
        "operationId": "predict_battery_health_api_v1_battery_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatteryTelemetryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatteryPredictionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/battery/rul": {
      "post": {
        "tags": [
          "Battery APM"
        ],
        "summary": "Predict Battery Rul",
        "description": "Predict Remaining Useful Life (RUL) only, using State of Health as input feature.",
        "operationId": "predict_battery_rul_api_v1_battery_rul_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatteryRULInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatteryRULOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/carbon/analysis": {
      "post": {
        "tags": [
          "Carbon Intelligence Tracker"
        ],
        "summary": "Analyze Vehicle Carbon",
        "description": "Perform a comparative carbon emissions analysis for a specific vehicle (baseline vs EV scenario)\nover a custom annual distance and fuel type.",
        "operationId": "analyze_vehicle_carbon_api_v1_carbon_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarbonAnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarbonAnalysisOutput"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/query": {
      "post": {
        "tags": [
          "AI Fleet Advisor"
        ],
        "summary": "Query Fleet Advisor",
        "description": "Query the AI Fleet Advisor to get insights about battery telemetry, \nreadiness schedules, or carbon calculations.",
        "operationId": "query_fleet_advisor_api_v1_chat_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatQueryInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatteryPredictionOutput": {
        "properties": {
          "battery_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Id",
            "description": "Identifier of the battery"
          },
          "state_of_health": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "State Of Health",
            "description": "Predicted State of Health (SOH) index [0-1]"
          },
          "remaining_useful_life_cycles": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Remaining Useful Life Cycles",
            "description": "Estimated remaining useful cycles before retirement"
          },
          "health_zone": {
            "type": "string",
            "title": "Health Zone",
            "description": "Categorized health zone (Healthy, Attention Required, Critical Failure Danger)"
          },
          "soh_model_used": {
            "type": "string",
            "title": "Soh Model Used",
            "description": "Source of SOH model prediction (trained/fallback_math)"
          },
          "rul_model_used": {
            "type": "string",
            "title": "Rul Model Used",
            "description": "Source of RUL model prediction (trained/fallback_math)"
          }
        },
        "type": "object",
        "required": [
          "state_of_health",
          "remaining_useful_life_cycles",
          "health_zone",
          "soh_model_used",
          "rul_model_used"
        ],
        "title": "BatteryPredictionOutput"
      },
      "BatteryRULInput": {
        "properties": {
          "cycle_number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Cycle Number",
            "description": "Current charge/discharge cycle count"
          },
          "voltage_v": {
            "type": "number",
            "minimum": 0.0,
            "title": "Voltage V",
            "description": "Measured terminal voltage in volts"
          },
          "temperature_c": {
            "type": "number",
            "title": "Temperature C",
            "description": "Core cell temperature in Celsius"
          },
          "capacity_ah": {
            "type": "number",
            "minimum": 0.0,
            "title": "Capacity Ah",
            "description": "Measured cell capacity in Ampere-hours"
          },
          "voltage_sag_v": {
            "type": "number",
            "title": "Voltage Sag V",
            "description": "Measured voltage sag under load in volts",
            "default": 0.0
          },
          "degradation_rate": {
            "type": "number",
            "title": "Degradation Rate",
            "description": "Calculated rate of degradation",
            "default": 0.0
          },
          "cycle_normalized": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cycle Normalized",
            "description": "Normalized cycle index"
          },
          "state_of_health": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "State Of Health",
            "description": "Estimated or calculated SOH"
          }
        },
        "type": "object",
        "required": [
          "cycle_number",
          "voltage_v",
          "temperature_c",
          "capacity_ah",
          "state_of_health"
        ],
        "title": "BatteryRULInput",
        "example": {
          "capacity_ah": 1.6,
          "cycle_normalized": 0.35,
          "cycle_number": 60,
          "degradation_rate": -0.003,
          "state_of_health": 0.88,
          "temperature_c": 28.0,
          "voltage_sag_v": 0.04,
          "voltage_v": 3.5
        }
      },
      "BatteryRULOutput": {
        "properties": {
          "remaining_useful_life_cycles": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Remaining Useful Life Cycles",
            "description": "Estimated remaining cycles before retirement"
          },
          "rul_model_used": {
            "type": "string",
            "title": "Rul Model Used",
            "description": "Source of model prediction"
          }
        },
        "type": "object",
        "required": [
          "remaining_useful_life_cycles",
          "rul_model_used"
        ],
        "title": "BatteryRULOutput"
      },
      "BatteryTelemetryInput": {
        "properties": {
          "cycle_number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Cycle Number",
            "description": "Current charge/discharge cycle count"
          },
          "voltage_v": {
            "type": "number",
            "minimum": 0.0,
            "title": "Voltage V",
            "description": "Measured terminal voltage in volts"
          },
          "temperature_c": {
            "type": "number",
            "title": "Temperature C",
            "description": "Core cell temperature in Celsius"
          },
          "capacity_ah": {
            "type": "number",
            "minimum": 0.0,
            "title": "Capacity Ah",
            "description": "Measured cell capacity in Ampere-hours"
          },
          "voltage_sag_v": {
            "type": "number",
            "title": "Voltage Sag V",
            "description": "Measured voltage sag under load in volts",
            "default": 0.0
          },
          "degradation_rate": {
            "type": "number",
            "title": "Degradation Rate",
            "description": "Calculated rate of degradation",
            "default": 0.0
          },
          "cycle_normalized": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cycle Normalized",
            "description": "Normalized cycle index (cycle/max_cycles)"
          }
        },
        "type": "object",
        "required": [
          "cycle_number",
          "voltage_v",
          "temperature_c",
          "capacity_ah"
        ],
        "title": "BatteryTelemetryInput",
        "example": {
          "capacity_ah": 1.75,
          "cycle_normalized": 0.3,
          "cycle_number": 50,
          "degradation_rate": -0.002,
          "temperature_c": 32.5,
          "voltage_sag_v": 0.03,
          "voltage_v": 3.65
        }
      },
      "CarbonAnalysisInput": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id",
            "description": "Vehicle ID to evaluate"
          },
          "annual_distance_km": {
            "type": "number",
            "minimum": 0.0,
            "title": "Annual Distance Km",
            "description": "Annual distance travelled in km"
          },
          "fuel_type": {
            "type": "string",
            "title": "Fuel Type",
            "description": "Fuel type (Diesel/Petrol/CNG)",
            "default": "Diesel"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "annual_distance_km"
        ],
        "title": "CarbonAnalysisInput",
        "example": {
          "annual_distance_km": 15000.0,
          "fuel_type": "Diesel",
          "vehicle_id": "VH-0013"
        }
      },
      "CarbonAnalysisOutput": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id",
            "description": "Vehicle ID analyzed"
          },
          "baseline_annual_co2_kg": {
            "type": "number",
            "title": "Baseline Annual Co2 Kg",
            "description": "Estimated ICE baseline CO2 emissions per year"
          },
          "ev_projected_co2_kg": {
            "type": "number",
            "title": "Ev Projected Co2 Kg",
            "description": "Projected grid/EV charging emissions per year"
          },
          "net_annual_savings_kg": {
            "type": "number",
            "title": "Net Annual Savings Kg",
            "description": "Net annual saved carbon in kg"
          },
          "savings_percentage": {
            "type": "number",
            "title": "Savings Percentage",
            "description": "Emissions reduction efficiency percentage"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "baseline_annual_co2_kg",
          "ev_projected_co2_kg",
          "net_annual_savings_kg",
          "savings_percentage"
        ],
        "title": "CarbonAnalysisOutput"
      },
...
      "ChatQueryInput": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "User query relating to fleet performance, battery health, or carbon indices"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Optional telemetry parameters to enhance prompt context"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatQueryInput"
      },
      "ChatQueryResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response",
            "description": "Generated markdown answer from the AI fleet advisor"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "Referenced datasets or telemetry records",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "response"
        ],
        "title": "ChatQueryResponse"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Detailed description of the error"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique UUID associated with the request log"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "example": {
          "detail": "Requested resource was not found.",
          "request_id": "8f8c7b8c-5544-4321-9abc-def012345678"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}